requirements.txt
# 核心依赖
PyYAML>=6.0
openai>=1.0.0  # 用于 openai（OpenAI 兼容渠道）
google-genai>=0.5.0  # 用于 gemini（Gemini 新版 SDK）
requests>=2.28.0

# 可选依赖（推荐安装以获得更好的体验）
python-dotenv>=1.0.0  # 更健壮的 .env 文件解析
typing-extensions>=4.0.0; python_version < "3.8"  # Python 3.7 的类型支持
